refactor(model): Eliminate embedding of Document :bug:
I've eliminated the embedding advice to use with the Documenter
interface. The type implementing the Documenter interface should have an
ID, CreatedOn and UpdatedOn attributes, with the mgo layout for naming
the attributes. I've removed the Document type since it would'n be
necessary for model package, nor handler. I still need to point this on
the documentation.