fix(example): Change handler to pass on real test of Find :bug:
I've changed the handler package code, on the Find method for the Handle
object. It makes some encoding and decoding, to use a bson.M version of
the model.Documenter received as a search object. It also convers the
bson.M object received to a model.Documenter object using Unmarshal.